consolidate prismatic-spray/hello-node/all-in-one (fixes #2298) (#2378)
This change consolidates various editions of prismatic spray into all-in-one, and just names them "client" and "server". It incidentally also adds an open-source-ready dockerfile, should we ever decide to take this stuff public. I was too far along with this change to throw it away, and I think it mostly makes sense even if we never OSS.

Notes:

- I reworked the directory structure someone, to consolidate similar files together.
- /usr/bin/nodejs is already installed on the build image, so it can just be copied as an #@ARTIFACT rather than re-installed. this removes the need for `all-on-one-base`.
- I added Dockerfiles for both the client and the server; my thinking is that when/if folks decide to move to a "real" installation, they may want to take their demo client and server with them. Now they can! Still needs to be documented, probably in the guides. (cc @brookshelley)

The new directory layout for all-in-one is:
```
bin/
  check-time.sh
  init.sh
  tbncollect.sh
client/
  create-workers.js
  Dockerfile
  index.html
  nginx.conf
  start.sh
config/
  clusters.yml
  nginx.conf
Dockerfile__oss__
Dockerfile__private__
README.md
server/
  Dockerfile
  server.js
  start-blue.sh
  start-green.sh
  start-yellow.sh
```

fbshipit-source-id: 28886bc9db0aa9151554ba282febb18bf9635e10